requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
Authlib==1.3.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.0
Deprecated==1.2.14
deprecation==2.1.0
fastapi==0.109.0
frozenlist==1.4.1
fusionauth-client==1.48.0
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
libsql-client==0.3.0
multidict==6.0.4
packaging==23.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.9
PyYAML==6.0.1
realtime==1.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
storage3==0.7.0
StrEnum==0.4.15
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==11.0.3
wrapt==1.16.0
yarl==1.9.4
